+++ src/webgpu/constants.js
@@ -8,3 +8,14 @@ export const TRANSMISSIVE_BACKGROUND_OVERLAY = 1;
 
 // Attenuate light based on transmitted
 export const TRANSMISSIVE_BACKGROUND_TRANSPARENT = 2;
+
+// Layout of the r32uint "sample count" target: two flag bits then a 30 bit count.
+
+// this pixel already has a ray on the queue
+export const SAMPLE_ACTIVE_FLAG = 0x80000000;
+
+// this pixel has had a camera ray at least once. pixels that never do, like the gaps between array
+// camera viewports, are skipped when tallying so they don't drag the min and average down
+export const SAMPLE_DISPATCHED_FLAG = 0x40000000;
+
+export const SAMPLE_COUNT_MASK = 0x3FFFFFFF;
